fix   一致性任务修改， 推送系统 60%

+++ taki-mall-market-api/src/main/java/com/taki/market/domain/dto/SaveOrUpdatePromotionDTO.java
@@ -34,5 +34,5 @@ public class SaveOrUpdatePromotionDTO {
     /**
      * 活动创建、修改人
      */
-    private Integer createUser;
+    private Long createUser;
 }
